+++ dali/test/python/checkpointing/test_dali_stateless_operators.py
@@ -24,6 +24,7 @@ from test_utils import (
     check_numba_compatibility_cpu,
     has_operator,
     restrict_platform,
+    is_of_supported,
 )
 from nose2.tools import params, cartesian_params
 from nose_utils import assert_raises, SkipTest, attr
@@ -575,8 +576,6 @@ def test_preemphasis_filter_stateless(device):
 
 @stateless_signed_off("optical_flow")
 def test_optical_flow_stateless():
-    from test_optical_flow import is_of_supported
-
     if not is_of_supported():
         raise SkipTest("Optical Flow is not supported on this platform")
     check_single_sequence_input(fn.optical_flow, "gpu")
