+++ dali/test/python/test_dali_variable_batch_size.py
@@ -29,13 +29,13 @@ from nvidia.dali.plugin.numba.fn.experimental import numba_function
 import test_utils
 from segmentation_test_utils import make_batch_select_masks
 from test_detection_pipeline import coco_anchors
-from test_optical_flow import load_frames, is_of_supported
 from test_utils import (
     module_functions,
     has_operator,
     restrict_platform,
     check_numba_compatibility_cpu,
     check_numba_compatibility_gpu,
+    is_of_supported,
 )
 
 """
@@ -1271,9 +1271,7 @@ def test_optical_flow():
         pipe.set_outputs(processed)
         return pipe
 
-    max_batch_size = 5
-    bach_sizes = [max_batch_size // 2, max_batch_size // 4, max_batch_size]
-    input_data = [[load_frames() for _ in range(bs)] for bs in bach_sizes]
+    input_data = generate_data(5, 2, (10, 160, 80, 3), lo=0, hi=255, dtype=np.uint8)
     check_pipeline(input_data, pipeline_fn=pipe, devices=["gpu"], input_layout="FHWC")
