+++ qa/TL0_self_test_Ampere/test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -ex
 
-pip_packages='${python_test_runner_package} numpy'
+pip_packages='${python_test_runner_package} numpy opencv-python nvidia-ml-py==11.450.51'
 
 target_dir=./dali/test/python
 
@@ -33,8 +33,9 @@ test_body() {
   ${python_new_invoke_test} -s decoder test_image
 
   # test Optical Flow
-  ${python_invoke_test} test_optical_flow.py
+  ${python_new_invoke_test} -s operator_1 test_optical_flow
   ${python_invoke_test} test_dali_variable_batch_size.py:test_optical_flow
+  ${python_invoke_test} test_dali_stateless_operators.py:test_optical_flow_stateless
 }
 
 pushd ../..
